requirements.txt
absl-py==0.13.0
AMFM-decompy==1.0.11
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==21.4.0
audioread==2.1.9
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
beautifulsoup4 @ file:///tmp/build/80754af9/beautifulsoup4_1631874778482/work
bleach==4.1.0
brotlipy==0.7.0
cachetools==4.2.2
certifi==2021.10.8
cffi @ file:///opt/conda/conda-bld/cffi_1642701102775/work
chardet @ file:///tmp/build/80754af9/chardet_1607706746162/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.0.3
conda==4.11.0
conda-build==3.21.4
conda-package-handling @ file:///tmp/build/80754af9/conda-package-handling_1618262148928/work
configparser==5.2.0
cryptography @ file:///tmp/build/80754af9/cryptography_1639400846433/work
cycler==0.10.0
debugpy==1.5.1
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
defusedxml==0.7.1
dnspython==2.1.0
docker-pycreds==0.4.0
entrypoints==0.4
executing==0.8.3
filelock @ file:///opt/conda/conda-bld/filelock_1642510437405/work
Flask==2.0.3
Flask-Cors==3.0.10
gitdb==4.0.9
GitPython==3.1.26
glob2 @ file:///home/linux1/recipes/ci/glob2_1610991677669/work
google-auth==1.35.0
google-auth-oauthlib==0.4.5
grpcio==1.39.0
hgtk==0.1.3
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
importlib-resources==5.4.0
ipykernel==6.9.1
ipython==8.1.1
ipython-genutils==0.2.0
ipywidgets==7.6.5
itsdangerous==2.1.0
jedi==0.17.0
Jinja2 @ file:///tmp/build/80754af9/jinja2_1635780242639/work
joblib==1.0.1
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.2
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
libarchive-c @ file:///tmp/build/80754af9/python-libarchive-c_1617780486945/work
librosa==0.8.1
llvmlite==0.37.0
Markdown==3.3.4
MarkupSafe @ file:///tmp/build/80754af9/markupsafe_1621528148836/work
matplotlib==3.4.3
matplotlib-inline==0.1.3
mido==1.2.9
mistune==0.8.4
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
multipledispatch==0.6.0
nano==0.10.0
nbclient==0.5.12
nbconvert==6.4.2
nbformat==5.1.3
nest-asyncio==1.5.4
notebook==6.4.8
numba==0.54.0
numpy==1.20.3
oauthlib==3.1.1
olefile @ file:///Users/ktietz/demo/mc3/conda-bld/olefile_1629805411829/work
packaging==21.0
pandocfilters==1.5.0
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pathtools==0.1.2
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==8.4.0
pkginfo @ file:///tmp/build/80754af9/pkginfo_1643162084911/work
pooch==1.5.1
prometheus-client==0.13.1
promise==2.3
prompt-toolkit @ file:///tmp/build/80754af9/prompt-toolkit_1633440160888/work
protobuf==3.17.3
psutil @ file:///tmp/build/80754af9/psutil_1612298023621/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycosat==0.6.3
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
Pygments @ file:///tmp/build/80754af9/pygments_1629234116488/work
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1635333100036/work
pyparsing==2.4.7
pyrsistent==0.18.1
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
pysptk==0.1.18
python-dateutil==2.8.2
python-etcd==0.4.5
pytz==2021.3
PyYAML==5.4.1
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.1
requests @ file:///opt/conda/conda-bld/requests_1641824580448/work
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.7.2
ruamel-yaml-conda @ file:///tmp/build/80754af9/ruamel_yaml_1616016699510/work
scikit-learn==0.24.2
scipy==1.7.1
Send2Trash==1.8.0
sentry-sdk==1.5.4
shortuuid==1.0.8
six @ file:///tmp/build/80754af9/six_1623709665295/work
smmap==5.0.0
SoundFile==0.10.3.post1
soupsieve @ file:///tmp/build/80754af9/soupsieve_1636706018808/work
stack-data==0.2.0
subprocess32==3.5.4
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
termcolor==1.1.0
terminado==0.13.2
testpath==0.6.0
threadpoolctl==2.2.0
torch==1.10.2
torchaudio==0.10.2
torchelastic==0.2.2
torchtext==0.11.2
torchvision==0.11.3
tornado==6.1
tqdm @ file:///tmp/build/80754af9/tqdm_1635330843403/work
traitlets @ file:///tmp/build/80754af9/traitlets_1636710298902/work
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1631814937681/work
urllib3 @ file:///opt/conda/conda-bld/urllib3_1643638302206/work
wandb==0.12.9
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.2
yaspin==2.1.0
zipp==3.7.0
